+++ obj/src/services/CloudFunctionService.js
@@ -187,10 +187,11 @@ class CloudFunctionService {
             // Validate object
             if (schema && req) {
                 // Perform validation
+                let params = Object.assign({}, req.params, req.query, { body: req.body });
                 let correlationId = this.getCorrelationId(req);
-                let err = schema.validateAndReturnException(correlationId, req.body, false);
+                let err = schema.validateAndReturnException(correlationId, params, false);
                 if (err) {
-                    throw err;
+                    pip_services3_rpc_nodex_1.HttpResponseSender.sendError(req, res, err);
                 }
             }
             return action.call(this, req, res);