+++ obj/test/CloudFunctionRequestSchema.d.ts
@@ -1,4 +1,0 @@
-import { ObjectSchema } from 'pip-services3-commons-nodex';
-export declare class CloudFunctionRequestSchema extends ObjectSchema {
-    constructor();
-}