+++ test/containers/DummyCloudFunction.ts
@@ -88,32 +88,47 @@ export class DummyCloudFunction extends CloudFunction {
         this.registerAction(
             'get_dummies',
             new ObjectSchema(true)
-                .withOptionalProperty("filter", new FilterParamsSchema())
-                .withOptionalProperty("paging", new PagingParamsSchema())
+                .withOptionalProperty('body',
+                    new ObjectSchema(true)
+                        .withOptionalProperty("filter", new FilterParamsSchema())
+                        .withOptionalProperty("paging", new PagingParamsSchema())
+                )
             , this.getPageByFilter);
 
         this.registerAction(
             'get_dummy_by_id',
             new ObjectSchema(true)
-                .withOptionalProperty("dummy_id", TypeCode.String)
+                .withOptionalProperty("body",
+                    new ObjectSchema(true)
+                        .withOptionalProperty("dummy_id", TypeCode.String)
+                )
             , this.getOneById);
 
         this.registerAction(
             'create_dummy',
             new ObjectSchema(true)
-                .withRequiredProperty("dummy", new DummySchema())
+                .withOptionalProperty("body",
+                    new ObjectSchema(true)
+                        .withRequiredProperty("dummy", new DummySchema())
+                )
             , this.create);
 
         this.registerAction(
             'update_dummy',
             new ObjectSchema(true)
-                .withRequiredProperty("dummy", new DummySchema())
+                .withOptionalProperty("body",
+                    new ObjectSchema(true)
+                        .withRequiredProperty("dummy", new DummySchema())
+                )
             , this.update);
 
         this.registerAction(
             'delete_dummy',
             new ObjectSchema(true)
-                .withOptionalProperty("dummy_id", TypeCode.String)
+                .withOptionalProperty("body",
+                    new ObjectSchema(true)
+                        .withOptionalProperty("dummy_id", TypeCode.String)
+                )
             , this.deleteById);
     }
 }