+++ test/containers/DummyCloudFunctionFixture.ts
@@ -115,5 +115,14 @@ export class DummyCloudFunctionFixture {
             dummy_id: dummy1.id
         });
         assert.isNull(dummy || null);
+
+
+        // Failed validation
+        // let err = await this.httpInvoke({
+        //     cmd: 'create_dummy',
+        //     dummy: null
+        // })
+
+        // assert.equal(err.restCode, 'INVALID_DATA');
     }
 }