+++ test/services/DummyCloudFunctionFixture.ts
@@ -114,5 +114,13 @@ export class DummyCloudFunctionFixture {
             dummy_id: dummy1.id
         });
         assert.isNull(dummy || null);
+
+        // Failed validation
+        let err = await this.httpInvoke({
+            cmd: 'dummies.create_dummy',
+            dummy: null
+        })
+
+        assert.equal(err.restCode, 'INVALID_DATA');
     }
 }