+++ CHANGELOG.md
@@ -1,11 +1,13 @@
 # <img src="https://uploads-ssl.webflow.com/5ea5d3315186cf5ec60c3ee4/5edf1c94ce4c859f2b188094_logo.svg" alt="Pip.Services Logo" width="200">
 <br/> Google Cloud Platform specific components for Node.js / ES2017 Changelog
 
-## <a name="1.0.1-1.0.3"></a> 1.0.1-1.0.3 (2022-03-09)
+## <a name="1.0.1-1.0.4"></a> 1.0.1-1.0.4 (2022-03-09)
 
 ### Bug Fixes
-- Fixed imports, update docs
-- Fixed error processing for requests
+- Fixed imports 
+- Fixed schemas validation
+- Fixed error processing
+- Fixed docs
 
 
 ## <a name="1.0.0"></a> 1.0.0 (2022-02-15)